// src/service-api/generated/models/importLeadRow.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */
import type { ImportLeadRowRândSursă } from './importLeadRowRândSursă';
import type { ImportLeadRowSourceRow } from './importLeadRowSourceRow';
import type { ImportLeadRowHasItTeam } from './importLeadRowHasItTeam';

export interface ImportLeadRow {
  'Status contact'?: string;
  'Rând sursă'?: ImportLeadRowRândSursă;
  Note?: string;
  Website?: string;
  'Echipă IT'?: string;
  'Pain point'?: string;
  LinkedIn?: string;
  Email?: string;
  Telefon?: string;
  Companie?: string;
  Functie?: string;
  Funcție?: string;
  Nume?: string;
  'Canal principal'?: string;
  Industrie?: string;
  'Pachet principal'?: string;
  Prioritate?: string;
  contactStatus?: string;
  /** @nullable */
  sourceRow?: ImportLeadRowSourceRow;
  notes?: string;
  website?: string;
  /** @nullable */
  hasItTeam?: ImportLeadRowHasItTeam;
  painPoint?: string;
  linkedinUrl?: string;
  email?: string;
  phone?: string;
  companyName?: string;
  jobTitle?: string;
  contactName?: string;
  channel?: string;
  industry?: string;
  packageName?: string;
  priority?: string;
}